# backend/openapi_temp.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZiggyAI Trading Platform",
    "description": "Complete AI-powered trading platform with market analysis, news, signals, and execution",
    "version": "1.0.0"
  },
  "paths": {
...
    "/feedback/feedback/decision": {
      "post": {
        "tags": ["feedback", "feedback"],
        "summary": "Submit Decision Feedback",
        "description": "Submit feedback on a trading decision.\n\nArgs:\n    feedback: Feedback data including event_id, rating, tags, and note\n\nReturns:\n    Feedback submission response",
        "operationId": "submit_decision_feedback_feedback_feedback_decision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FeedbackRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FeedbackResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/feedback/feedback/event/{event_id}": {
      "get": {
        "tags": ["feedback", "feedback"],
        "summary": "Get Event Feedback",
        "description": "Get feedback summary for a specific event.\n\nArgs:\n    event_id: Event ID to get feedback for\n\nReturns:\n    Feedback summary including counts and recent feedback",
        "operationId": "get_event_feedback_feedback_feedback_event__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Event Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FeedbackSummary" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/feedback/feedback/bulk": {
      "post": {
        "tags": ["feedback", "feedback"],
        "summary": "Submit Bulk Feedback",
        "description": "Submit feedback for multiple decisions at once.\n\nArgs:\n    request: Bulk feedback request with multiple feedback items\n\nReturns:\n    Bulk submission results",
        "operationId": "submit_bulk_feedback_feedback_feedback_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkFeedbackRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkFeedbackRequest": {
        "properties": {
          "feedback_items": {
            "items": { "$ref": "#/components/schemas/FeedbackRequest" },
            "type": "array",
            "title": "Feedback Items",
            "description": "List of feedback items"
          }
        },
        "type": "object",
        "required": ["feedback_items"],
        "title": "BulkFeedbackRequest",
        "description": "Request model for bulk feedback submission.",
        "example": {
          "feedback_items": [
            {
              "event_id": "event1",
              "note": "Great call",
              "rating": "GOOD",
              "tags": ["accurate"]
            },
            {
              "event_id": "event2",
              "note": "Missed the mark",
              "rating": "BAD",
              "tags": ["poor_timing"]
            }
          ]
        }
      },
      "BulkFeedbackResponse": {
        "properties": {
          "total_submitted": { "type": "integer", "title": "Total Submitted" },
          "successful": { "type": "integer", "title": "Successful" },
          "failed": { "type": "integer", "title": "Failed" },
          "feedback_ids": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Feedback Ids"
          },
          "errors": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "total_submitted",
          "successful",
          "failed",
          "feedback_ids",
          "errors"
        ],
        "title": "BulkFeedbackResponse",
        "description": "Response model for bulk feedback submission."
      },
...
      "FeedbackRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id",
            "description": "ID of the event to provide feedback on"
          },
          "rating": {
            "type": "string",
            "title": "Rating",
            "description": "Feedback rating: GOOD or BAD"
          },
          "tags": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Tags",
            "description": "Optional feedback tags"
          },
          "note": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Note",
            "description": "Optional feedback note"
          }
        },
        "type": "object",
        "required": ["event_id", "rating"],
        "title": "FeedbackRequest",
        "description": "Request model for decision feedback.",
        "example": {
          "event_id": "12345678-1234-1234-1234-123456789012",
          "note": "Model correctly predicted the earnings reaction",
          "rating": "GOOD",
          "tags": ["accurate_prediction", "good_timing"]
        }
      },
      "FeedbackResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "feedback_id": { "type": "string", "title": "Feedback Id" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["success", "feedback_id", "message"],
        "title": "FeedbackResponse",
        "description": "Response model for feedback submission."
      },
      "FeedbackSummary": {
        "properties": {
          "event_id": { "type": "string", "title": "Event Id" },
          "total_feedback_count": {
            "type": "integer",
            "title": "Total Feedback Count"
          },
          "rating_counts": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Rating Counts"
          },
          "recent_feedback": {
            "items": { "additionalProperties": true, "type": "object" },
            "type": "array",
            "title": "Recent Feedback"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "total_feedback_count",
          "rating_counts",
          "recent_feedback"
        ],
        "title": "FeedbackSummary",
        "description": "Summary of feedback for an event."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}